+++ SwiftTubeApp/Sources/SwiftTubeApp/PlaybackEngine.swift
@@ -10,7 +10,23 @@ struct MediaStreamRequest: Hashable, Sendable {
     }
 }
 
+enum MPVPlaybackMode: Hashable, Sendable {
+    case direct
+    case adaptiveManifest
+}
+
 struct MPVPlaybackRequest: Hashable, Sendable {
     let video: MediaStreamRequest
     let audio: MediaStreamRequest?
+    let mode: MPVPlaybackMode
+
+    init(
+        video: MediaStreamRequest,
+        audio: MediaStreamRequest?,
+        mode: MPVPlaybackMode = .direct
+    ) {
+        self.video = video
+        self.audio = audio
+        self.mode = mode
+    }
 }